@@ -49,7 +49,7 @@ using System.Windows;
 //可以指定所有这些值，也可以使用“生成号”和“修订号”的默认值
 //通过使用 "*"，如下所示:
 // [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion("1.1.1.26")]
-[assembly: AssemblyFileVersion("1.1.1.26")]
+[assembly: AssemblyVersion("1.1.1.27")]
+[assembly: AssemblyFileVersion("1.1.1.27")]
 
 [assembly: AssemblyInformationalVersion("1.1.1")]
