@@ -37,7 +37,7 @@
 
         private readonly IEventAggregator _eventAggregator;
 
-        private Encoding fallbackEncoding => Encoding.GetEncoding(_configurationService.Get<MiscConfiguration>().DicomEncoding);
+        private Encoding _fallbackEncoding => Encoding.GetEncoding(_configurationService.Get<MiscConfiguration>().DicomEncoding);
 
         private DicomFile _currentFile;
 
@@ -160,7 +160,7 @@
 
             DicomItems.Clear();
 
-            _currentFile = await DicomFile.OpenAsync(file, fallbackEncoding);
+            _currentFile = await DicomFile.OpenAsync(file, _fallbackEncoding);
             _currentFile.FileMetaInfo.NotValidated();
             _currentFile.Dataset.NotValidated();
 
@@ -336,7 +336,7 @@
         private void AddOrUpdateDicomItem(DicomDataset dataset, DicomVR vr, DicomTag tag, string[] values)
         {
             Encoding encoding = _currentFile.Dataset.TryGetValue(DicomTag.SpecificCharacterSet, 0, out string charset)
-                ? DicomEncoding.GetEncoding(charset) : fallbackEncoding;
+                ? DicomEncoding.GetEncoding(charset) : _fallbackEncoding;
 
             if (vr == DicomVR.OB || vr == DicomVR.UN)
             {
